AlgorithmExecutor + test_post_request file rename
Also, remove .flake8
