--- mipengine/controller/algorithm_executor/.flake8
@@ -1,2 +1,0 @@
-[flake8]
-ignore = E501,
